product/customer/requirements.txt
flask==3.0.1
flask-sqlalchemy==3.1.1
flask-migrate==2.7.0
flask-jwt-extended==4.6.0
flask_script==2.0.6
sqlalchemy_utils==0.41.1
Werkzeug==3.0.1
Flask-CLI==0.4.0
pymysql==1.1.0
mysqlclient==2.2.1
cryptography==42.0.1